+++ cmd/chain/time/process.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	eth2client "github.com/attestantio/go-eth2-client"
-	spec "github.com/attestantio/go-eth2-client/spec/phase0"
+	"github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/pkg/errors"
 	"github.com/wealdtech/ethdo/util"
 )
@@ -41,6 +41,7 @@ func process(ctx context.Context, data *dataIn) (*dataOut, error) {
 
 	slotsPerEpoch := config["SLOTS_PER_EPOCH"].(uint64)
 	slotDuration := config["SECONDS_PER_SLOT"].(time.Duration)
+	epochsPerSyncCommitteePeriod := config["EPOCHS_PER_SYNC_COMMITTEE_PERIOD"].(uint64)
 	genesis, err := eth2Client.(eth2client.GenesisProvider).Genesis(ctx)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to obtain genesis data")
@@ -59,13 +60,13 @@ func process(ctx context.Context, data *dataIn) (*dataOut, error) {
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to parse slot")
 		}
-		results.slot = spec.Slot(slot)
+		results.slot = phase0.Slot(slot)
 	case data.epoch != "":
 		epoch, err := strconv.ParseUint(data.epoch, 10, 64)
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to parse epoch")
 		}
-		results.slot = spec.Slot(epoch * slotsPerEpoch)
+		results.slot = phase0.Slot(epoch * slotsPerEpoch)
 	case data.timestamp != "":
 		timestamp, err := time.Parse("2006-01-02T15:04:05-0700", data.timestamp)
 		if err != nil {
@@ -75,15 +76,20 @@ func process(ctx context.Context, data *dataIn) (*dataOut, error) {
 		if secs < 0 {
 			return nil, errors.New("timestamp prior to genesis")
 		}
-		results.slot = spec.Slot(secs / slotDuration)
+		results.slot = phase0.Slot(secs / slotDuration)
 	}
 
 	// Fill in the info given the slot.
 	results.slotStart = genesis.GenesisTime.Add(time.Duration(results.slot) * slotDuration)
 	results.slotEnd = genesis.GenesisTime.Add(time.Duration(results.slot+1) * slotDuration)
-	results.epoch = spec.Epoch(uint64(results.slot) / slotsPerEpoch)
+	results.epoch = phase0.Epoch(uint64(results.slot) / slotsPerEpoch)
 	results.epochStart = genesis.GenesisTime.Add(time.Duration(uint64(results.epoch)*slotsPerEpoch) * slotDuration)
 	results.epochEnd = genesis.GenesisTime.Add(time.Duration(uint64(results.epoch+1)*slotsPerEpoch) * slotDuration)
+	results.syncCommitteePeriod = uint64(results.epoch) / epochsPerSyncCommitteePeriod
+	results.syncCommitteePeriodEpochStart = phase0.Epoch(results.syncCommitteePeriod * epochsPerSyncCommitteePeriod)
+	results.syncCommitteePeriodEpochEnd = phase0.Epoch((results.syncCommitteePeriod+1)*epochsPerSyncCommitteePeriod) - 1
+	results.syncCommitteePeriodStart = genesis.GenesisTime.Add(time.Duration(uint64(results.syncCommitteePeriodEpochStart)*slotsPerEpoch) * slotDuration)
+	results.syncCommitteePeriodEnd = genesis.GenesisTime.Add(time.Duration(uint64(results.syncCommitteePeriodEpochEnd)*slotsPerEpoch) * slotDuration)
 
 	return results, nil
 }